Support camera clip editing
